requirements.txt
mysql-connector-python>=9.0.0
reportlab>=4.2.5
tkcalendar>=1.6.1